LWDEV-3468: get accounts by margin usage level, close positions by ac…

+++ src/MarginTrading.Common/BackendContracts/MarginTradingAccountBackendContract.cs
@@ -18,7 +18,7 @@ namespace MarginTrading.Common.BackendContracts
         public decimal UsedMargin { get; set; }
         public decimal MarginInit { get; set; }
         public decimal PnL { get; set; }
-        public decimal OpenPositionsCount { get; set; }
+        public int OpenPositionsCount { get; set; }
         public decimal MarginUsageLevel { get; set; }
         public bool IsLive { get; set; }
     }
